Implement #if and friends.
With this change, tests 41 through 49 all pass. (The defined operator
appears to be somehow broken so that test 50 doesn't pass yet.)

+++ glcpp.h
@@ -49,7 +49,7 @@ typedef struct token_list token_list_t;
 
 typedef union YYSTYPE
 {
-	int ival;
+	intmax_t ival;
 	char *str;
 	string_list_t *string_list;
 	token_t *token;
